chag tag: added --cleanup=whitespace (fixes #7)
chag tag caused headings that are part of a changelog entry content
to be ignored because git interpreted them as comments.

This fix preserves lines prefixed with a # when creating the annotated tag and
therefore all lines that are markdown headers are preserved in the annotation
message.

This patch also fixes the tests according to the new behavior.